Rename hgLog library file
We're about to add more code to this, so it should have a more generic
name to reflect what it's going to contain.
