+++ src/lib/components/HgLogGraph.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import SvgRailroadGraph from './SvgRailroadGraph.svelte'
-    import { buildParentData } from '$lib/log/parentDataHgLog'
+    import { buildParentData } from '$lib/log/hgLog'
 
     export let hglog: Record<string, any>[]
     export let tableBody: HTMLTableSectionElement
